% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoshuaIDDproj4-package.R
\docType{package}
\name{JoshuaIDDproj4-package}
\alias{JoshuaIDDproj4}
\alias{JoshuaIDDproj4-package}
\title{JoshuaIDDproj4: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NiyeloStar1/JoshuaIDDproj4}
  \item Report bugs at \url{https://github.com/NiyeloStar1/JoshuaIDDproj4/issues}
}

}
\author{
\strong{Maintainer}: Gbeminiyi Oyedele \email{joshuagbeminiyi@gmail.com} (\href{https://orcid.org/0000000000}{ORCID}) [copyright holder]

}
\keyword{internal}
